requirements.txt
torch>=2.1
transformers>=4.45
accelerate>=0.33
huggingface_hub>=0.24
safetensors>=0.4
llm-siren>=0.1.0
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
tqdm>=4.66
pyyaml>=6.0
matplotlib>=3.8
scipy>=1.11

datasets>=2.20
